console-chat-bot/requirements.txt
openai==1.54.4
langchain==0.3.7
langchain-community==0.3.7
python-dotenv==1.0.1